\name{open_dev}
\alias{open_dev}
\title{Open a device from the filename extension}
\usage{
open_dev(filename, type = "cairo", ...)
}
\arguments{
  \item{filename}{filename to open}

  \item{type}{Type of device to open}

  \item{...}{arguments to be passed to device}
}
\value{
NULL, with deviced opened
}
\description{
Open a device from the filename extension
}
\keyword{viewer}

