\name{view.png}
\alias{view.png}
\title{View png in R}
\usage{
view.png(filename, viewer = "display", bg = TRUE)
}
\arguments{
  \item{filename}{filename of png}

  \item{viewer}{name of viewer in PATH}

  \item{bg}{run as background process}
}
\value{
Result from system
}
\description{
View a png in external viewer from R
}
\keyword{viewer}

