% Generated by roxygen2 (4.0.2): do not edit by hand
\name{open_dev}
\alias{open_dev}
\title{Open a device from the filename extension}
\usage{
open_dev(filename, type = "cairo", ...)
}
\arguments{
\item{filename}{filename to open}

\item{type}{Type of device to open, i.e. "cairo",
(see \code{\link{png}})}

\item{...}{arguments to be passed to device}
}
\value{
NULL, with deviced opened
}
\description{
Open a device from the filename extension.
This function allows you to simply open a device by using
the filename.  So that if you have .PNG as the extension it opens
a PNG, etc.
}
\examples{
fname = tempfile(fileext= ".png")
open_dev(fname)
plot(0, 0)
dev.off()
# view(fname)
# view(fname, viewer="open")
fname = tempfile(fileext= ".pdf")
open_dev(fname,height=7, width=7)
plot(0, 0)
dev.off()
# view(fname)
}
\keyword{viewer}

