% Generated by roxygen2 (4.0.1): do not edit by hand
\name{smallpdf.off}
\alias{pngpdf.off}
\alias{smallpdf.off}
\title{Close smallpdf device}
\usage{
smallpdf.off(pdfname, mypattern, dev, extra.opts = "-quality 100",
  outdir = tempdir(), clean = FALSE)
}
\arguments{
\item{pdfname}{filename of pdf to be created}

\item{mypattern}{regular expression pattern for device files
(from \code{\link{smallpdf}})}

\item{dev}{device used to open (from \code{\link{smallpdf}})}

\item{extra.opts}{Additional options sent to convert}

\item{outdir}{Directory of intermediate bitmap files (defaults to
tempdir())}

\item{clean}{Delete temporary bitmap files or not?}
}
\value{
The command for the conversion
}
\description{
Closes smallpdf device and converts the files into.  This is only worthwhile
when the pdf has a lot of information, such as when using \code{\link{image}}.
}
\examples{
\dontrun{
## not good for small plots
x = smallpdf(dev="jpeg")
dat = matrix(rnorm(1e3*4), ncol=4)
plot(dat[,1], dat[,2])
plot(dat[,3], dat[,4])
fname = file.path(tempdir(), "smallpdf.pdf")
smallpdf.off(pdfname = fname, mypattern=x$mypattern, dev=x$dev)
print(file.info(fname)$size)
# view.pdf(fname)
fname2 = file.path(tempdir(), "normalpdf.pdf")
pdf(fname2)
plot(dat[,1], dat[,2])
plot(dat[,3], dat[,4])
dev.off()
file.info(fname2)$size
print(file.info(fname2)$size / file.info(fname)$size)

## better for large plots
x = smallpdf(dev="jpeg")
dat = matrix(rnorm(1e6*4), ncol=4)
plot(dat[,1], dat[,2])
plot(dat[,3], dat[,4])
fname = file.path(tempdir(), "smallpdf.pdf")
smallpdf.off(pdfname = fname, mypattern=x$mypattern, dev=x$dev)
file.info(fname)$size
# [1] 3339399
# view.pdf(fname)
fname2 = file.path(tempdir(), "normalpdf.pdf")
pdf(fname2)
plot(dat[,1], dat[,2])
plot(dat[,3], dat[,4])
dev.off()
file.info(fname2)$size
# [1] 13648866
file.info(fname2)$size / file.info(fname)$size
# [1] 4.087222
}
}
\seealso{
dev.off
}

