\name{view.pdf}
\alias{view.pdf}
\title{View pdf from R}
\usage{
view.pdf(filename, viewer = getOption("pdfviewer"), bg = FALSE)
}
\arguments{
  \item{filename}{filename of pdf}

  \item{viewer}{<what param does>}

  \item{bg}{run as background process}
}
\value{
Result from system
}
\description{
View a pdf in an external viewer
}
\keyword{viewer}

