% Generated by roxygen2 (4.0.2): do not edit by hand
\name{view.pdf}
\alias{view.pdf}
\title{View PDF from R}
\usage{
view.pdf(filename, viewer = getOption("pdfviewer"), bg = FALSE)
}
\arguments{
\item{filename}{filename of PDF}

\item{viewer}{Name for the system command for viewer.  Defaults
to \code{pdfviewer} option}

\item{bg}{run as background process}
}
\value{
Result from \code{\link{system}} command
}
\description{
View a PDF in an external viewer
}
\keyword{viewer}

