% Generated by roxygen2 (4.0.1): do not edit by hand
\name{view.png}
\alias{view.png}
\title{View PNG in R}
\usage{
view.png(filename, viewer = "display", bg = FALSE)
}
\arguments{
\item{filename}{filename of PNG}

\item{viewer}{name of viewer in PATH.  Defaults
to "display" for ImageMagick}

\item{bg}{run as background process}
}
\value{
Result from \code{\link{system}} command
}
\description{
View a PNG in external viewer from R
}
\keyword{viewer}

